confirmation for deletion

+++ frontend/src/app/products/manage/manage.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import {Router} from '@angular/router';
 import {ProductsService, IProduct} from '../products.service';
 import {UserService} from 'src/app/user/user.service';
+import {MatDialog} from '@angular/material/dialog';
+import { DeleteComponent } from '../delete/delete.component';
 
 @Component({
   selector: 'app-products-manage',
@@ -16,7 +18,8 @@ export class ManageComponent implements OnInit {
   constructor(
     public userService: UserService,
     public router: Router,
-    private productService: ProductsService
+    private productService: ProductsService,
+    public dialog: MatDialog
   ) {
     this.reloadProducts();
   }
@@ -33,7 +36,11 @@ export class ManageComponent implements OnInit {
   }
 
   delete(product: IProduct): void {
-    this.productService.delete(product).subscribe( () => this.reloadProducts());
+    this.dialog.open(DeleteComponent).afterClosed().subscribe( result => {
+      if (result) {
+        this.productService.delete(product).subscribe( () => this.reloadProducts());
+      }
+    });
   }
 
   reloadProducts(): void {
